# OPT: parse_gitconfig_dump() origin path processing