Instead of doing this once per config item (many), do it once per path (few). On my system with 161 config items this leads to a 6x speed-up.

Before:

```
838 µs ± 8.54 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```

After:

```
142 µs ± 7.69 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```

Code:

```py
>>> import subprocess; dump=subprocess.run(['git', 'config', '-z', '--list', '--show-origin'], capture_output=True, text=True).stdout; from datalad.config import parse_gitconfig_dump
>>>  timeit p=parse_gitconfig_dump(dump)
```